requirements.txt
asn1crypto==0.24.0
bezier==0.9.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
cryptography==2.1.4
cycler==0.10.0
decorator==4.1.2
idna==2.8
matplotlib==2.1.1
numpy==1.15.4
Pillow==5.1.0
pkg-resources==0.0.0
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
PyNaCl==1.1.2
PyOpenGL==3.1.0
pyparsing==2.2.0
pyqtgraph==0.10.0
python-dateutil==2.7.5
pytz==2018.9
scipy==0.19.1
six==1.12.0
svgpathtools==1.3.3
svgwrite==1.2.1
